# doc/contracts/backend-api-v1.openapi.yaml
openapi: 3.1.0
info:
  title: LabRat Backend API
  version: 1.0.0
  description: >-
    Contract-first API for the NestJS modular monolith. All documented
    operations have closed DTOs and authoritative v1 behavior.
    Public Guest identities are restricted to one demo project with read/export
    access. Authenticated Guest writes, AI requests and invitation redemption
    return 403 public_guest_read_only; only GET/HEAD and POST logout are allowed.
    Guest request limits may return 429 too_many_requests on authenticated reads.
servers:
  - url: http://127.0.0.1:8788
    description: Local v1 development server
...
security:
  - cookieAuth: []
paths:
...
  /api/v1/projects/{projectId}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      tags: [Tenancy]
      operationId: getProject
      x-labrat-capability: read
      responses:
        "200":
          description: Full or minimum authorized Project shell.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Tenancy]
      operationId: updateProject
      x-labrat-capability: propose
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectRequest"
      responses:
        "200":
          description: Project updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/v1/projects/{projectId}/region-extraction-templates:
    get:
      tags:
        - Evidence
      operationId: listRegionExtractionTemplates
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/OpaqueId"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
        - name: includeArchived
          in: query
          schema:
            type: string
            enum:
              - "true"
              - "false"
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegionExtractionTemplatePage"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    post:
      tags:
        - Evidence
      operationId: createRegionExtractionTemplate
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/OpaqueId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRegionExtractionTemplateRequest"
      responses:
        "201":
          description: Success.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegionExtractionTemplateDetail"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/region-extraction-templates/{templateId}:
    get:
      tags:
        - Evidence
      operationId: getRegionExtractionTemplate
      parameters:
        - name: templateId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/OpaqueId"
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegionExtractionTemplateDetail"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/region-extraction-templates/{templateId}/versions:
    post:
      tags:
        - Evidence
      operationId: createRegionExtractionTemplateVersion
      parameters:
        - name: templateId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/OpaqueId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegionTemplateVersionRequest"
      responses:
        "201":
          description: Success.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegionExtractionTemplateDetail"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/region-extraction-templates/{templateId}/archive:
    post:
      tags:
        - Evidence
      operationId: archiveRegionExtractionTemplate
      parameters:
        - name: templateId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/OpaqueId"
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegionExtractionTemplateResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/region-extraction-template-versions/{versionId}/matches:
    post:
      tags:
        - Evidence
      operationId: matchRegionExtractionTemplate
      parameters:
        - name: versionId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/OpaqueId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MatchRegionTemplateRequest"
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegionTemplateMatches"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/region-extraction-template-versions/{versionId}/apply:
    post:
      tags:
        - Evidence
      operationId: applyRegionExtractionTemplate
      parameters:
        - name: versionId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/OpaqueId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplyRegionTemplateRequest"
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegionTemplateApplyResult"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /api/v1/projects/{projectId}/workbook-review-regions/confirm-batch:
    post:
      tags:
        - Evidence
      operationId: confirmWorkbookReviewRegionsBatch
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/OpaqueId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfirmRegionBatchRequest"
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegionBatchConfirmResult"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: labrat_session
  parameters:
...
    ProjectId:
      name: projectId
      in: path
      required: true
      schema: { $ref: "#/components/schemas/OpaqueId" }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 1, maxLength: 200, pattern: "^[A-Za-z0-9._:-]+$" }
    Cursor:
      name: cursor
      in: query
      required: false
      schema: { type: string, maxLength: 500 }
    Limit:
      name: limit
      in: query
      required: false
      schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
...
  responses:
    Unauthorized:
      description: Authentication is missing or expired.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Forbidden:
      description: The actor lacks the required capability.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: Resource not found in the authorized scope.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Conflict:
      description: Version, state, stale-head or idempotency conflict.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    ValidationError:
      description: Request failed bounded validation.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
  schemas:
    CreateRegionExtractionTemplateRequest:
      type: object
      additionalProperties: false
      required:
        - name
        - regionId
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 120
        description:
          type: string
          maxLength: 1000
        regionId:
          type: string
          minLength: 1
          maxLength: 160
    RegionTemplateVersionRequest:
      type: object
      additionalProperties: false
      required:
        - regionId
      properties:
        regionId:
          type: string
          minLength: 1
          maxLength: 160
    MatchRegionTemplateRequest:
      type: object
      additionalProperties: false
      required:
        - sourceDocumentIds
      properties:
        sourceDocumentIds:
          type: array
          items:
            type: string
            minLength: 1
            maxLength: 160
          maxItems: 100
          minItems: 1
          uniqueItems: true
    ApplyRegionTemplateRequest:
      type: object
      additionalProperties: false
      required:
        - sourceDocumentIds
      properties:
        sourceDocumentIds:
          type: array
          items:
            type: string
            minLength: 1
            maxLength: 160
          maxItems: 100
          minItems: 1
          uniqueItems: true
        onlyStatuses:
          type: array
          description: >-
            Match statuses to prefill. Defaults to exact and shifted. formula_mismatch
            prefills a block whose cells hold text where the template expects values;
            such regions carry a warning and are confirmed individually, never in a batch.
          items:
            type: string
            enum:
              - exact
              - shifted
              - formula_mismatch
          maxItems: 3
          minItems: 1
          uniqueItems: true
    ConfirmRegionBatchItem:
      type: object
      additionalProperties: false
      required:
        - regionId
        - revisionId
        - expectedRegionVersion
      properties:
        regionId:
          type: string
          minLength: 1
          maxLength: 160
        revisionId:
          type: string
          minLength: 1
          maxLength: 160
        expectedRegionVersion:
          type: integer
          minimum: 1
        linkedExperimentId:
          type: string
          minLength: 1
          maxLength: 160
    ConfirmRegionBatchRequest:
      type: object
      additionalProperties: false
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/ConfirmRegionBatchItem"
          maxItems: 200
          minItems: 1
...
    RegionExtractionTemplatePage:
      type: object
      additionalProperties: false
      required:
        - items
        - nextCursor
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/JsonObject"
          maxItems: 100
        nextCursor:
          $ref: "#/components/schemas/NullableString"
    RegionExtractionTemplateDetail:
      type: object
      additionalProperties: false
      required:
        - regionExtractionTemplate
        - versions
      properties:
        regionExtractionTemplate:
          $ref: "#/components/schemas/JsonObject"
        versions:
          type: array
          items:
            $ref: "#/components/schemas/JsonObject"
        sourceRegionLink:
          oneOf:
            - $ref: "#/components/schemas/JsonObject"
            - type: "null"
    RegionExtractionTemplateResponse:
      type: object
      additionalProperties: false
      required:
        - regionExtractionTemplate
      properties:
        regionExtractionTemplate:
          $ref: "#/components/schemas/JsonObject"
    RegionTemplateMatches:
      type: object
      additionalProperties: false
      required:
        - schemaVersion
        - regionExtractionTemplateId
        - templateName
        - templateVersionId
        - templateVersion
        - matches
        - summary
      properties:
        schemaVersion:
          type: string
        regionExtractionTemplateId:
          type: string
          minLength: 1
          maxLength: 160
        templateName:
          type: string
          minLength: 1
          maxLength: 120
        templateVersionId:
          type: string
          minLength: 1
          maxLength: 160
        templateVersion:
          type: integer
        matches:
          type: array
          items:
            $ref: "#/components/schemas/JsonObject"
          maxItems: 100
        summary:
          $ref: "#/components/schemas/JsonObject"
    RegionTemplateApplyResult:
      type: object
      additionalProperties: false
      required:
        - schemaVersion
        - regionExtractionTemplateId
        - templateName
        - templateVersionId
        - templateVersion
        - applied
        - skipped
      properties:
        schemaVersion:
          type: string
        regionExtractionTemplateId:
          type: string
          minLength: 1
          maxLength: 160
        templateName:
          type: string
          minLength: 1
          maxLength: 120
        templateVersionId:
          type: string
          minLength: 1
          maxLength: 160
        templateVersion:
          type: integer
        applied:
          type: array
          items:
            $ref: "#/components/schemas/JsonObject"
          maxItems: 100
        skipped:
          type: array
          items:
            $ref: "#/components/schemas/JsonObject"
          maxItems: 100
    RegionBatchConfirmResult:
      type: object
      additionalProperties: false
      required:
        - schemaVersion
        - projectId
        - results
        - confirmedCount
        - rejectedCount
      properties:
        schemaVersion:
          type: string
        projectId:
          type: string
          minLength: 1
          maxLength: 160
        results:
          type: array
          items:
            $ref: "#/components/schemas/JsonObject"
          maxItems: 200
        confirmedCount:
          type: integer
        rejectedCount:
          type: integer
...
    OpaqueId:
      type: string
      minLength: 1
      maxLength: 200
...
    Capability:
      type: string
      enum: [read, propose, approve, export, manage_access]
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [code, message, requestId]
          properties:
            code: { type: string, minLength: 1, maxLength: 100 }
            message: { type: string, minLength: 1, maxLength: 1000 }
            requestId: { type: string, minLength: 1, maxLength: 200 }
            details:
              type: object
              additionalProperties: true
...
    ProjectProfile:
      type: object
      additionalProperties: false
      properties:
        schemaVersion: { type: string, const: labrat.projectProfile.v1 }
        researchGoal: { type: string, maxLength: 20000 }
        experimentBackground: { type: string, maxLength: 20000 }
        materials: { type: string, maxLength: 20000 }
        methods: { type: string, maxLength: 20000 }
        instruments: { type: string, maxLength: 20000 }
        analysisNotes: { type: string, maxLength: 20000 }
        tags:
          type: array
          maxItems: 100
          items: { type: string, maxLength: 200 }
    ProjectWorkflowSummary:
      type: object
      additionalProperties: false
      required: [publishedExperimentCount, chartSpecCount, chartStyleProfileCount, reusableChartTemplateCount]
      properties:
        publishedExperimentCount: { type: integer, minimum: 0 }
        chartSpecCount: { type: integer, minimum: 0 }
        chartStyleProfileCount: { type: integer, minimum: 0 }
        reusableChartTemplateCount: { type: integer, minimum: 0 }
    Project:
      type: object
      additionalProperties: false
      required: [id, labId, name, description, status]
      properties:
        id: { $ref: "#/components/schemas/OpaqueId" }
        labId: { $ref: "#/components/schemas/OpaqueId" }
        name: { type: string }
        description: { type: string }
        status: { type: string, enum: [active, archived] }
        projectProfile: { $ref: "#/components/schemas/ProjectProfile" }
        shellOnly: { type: boolean, default: false }
        capabilities:
          type: array
          uniqueItems: true
          items: { $ref: "#/components/schemas/Capability" }
        workflowSummary: { $ref: "#/components/schemas/ProjectWorkflowSummary" }
...
    UpdateProjectRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name: { type: string, minLength: 1, maxLength: 300 }
        description: { type: string, maxLength: 5000 }
        status: { type: string, enum: [active, archived] }
    ProjectResponse:
      type: object
      additionalProperties: false
      required: [project]
      properties:
        project: { $ref: "#/components/schemas/Project" }
...
    NullableString:
      oneOf:
        - type: string
        - type: "null"
...
    JsonValue:
      oneOf:
        - type: "null"
        - type: boolean
        - type: number
        - type: string
        - type: array
          items: { $ref: "#/components/schemas/JsonValue" }
        - type: object
          additionalProperties: { $ref: "#/components/schemas/JsonValue" }
    JsonObject:
      type: object
      additionalProperties: { $ref: "#/components/schemas/JsonValue" }
...